chore: Refactor Ilox.Scanner
Improve the context of the scanner; it is unnecessary to collect both
the lexeme (previously called `current`) and the `literal` (removed), as
the literal is either the same as the lexeme or a minor modification
(removing the leading quote for `:string` tokens).

Reorganized the code and renamed the functions to be easier to follow.

Fixed number parsing to include the passed value as the lexeme.