chore: Refactor Ilox.Parser
Move the Ilox.ParserError from `lib/ilox/errors.ex` to be declared right
before the parser itself. These are closely related.

Add a real context map for scanned token parsing. This will prevent
printing to stderr during parsing (at least unless there is some sort of
callback mechanism added), but parsing will continue and contain all
errors at the end. This caused a lot of variable *name changes*, but the
primary change was in head parsing (as well as a few context updates
which were transparent previously).

As with the refactoring of Ilox.Scanner, renamed a number of internal
functions to be clearer on the behaviour. This marks *explicitly* moving
from the names in _Crafting Interpreters_, so it will require a bit more
translation moving forward, but this feels much more Elixir-native now.